refactor: rename serverSentEventsServices
