+++ app/frontend/src/services/serverSentEventsServices.ts
